Got rid of 'period' variable. No longer needs call input. edited Config
Removed 'read_args' function, changed the start message, changed the
'snap_description' assignment to not include the variable 'period',
changed the number of snapshots to keep to a single variable (rid of
'period'), in the config.sample, deleted the 3 different 'period'
variable values and replaced with single value that is configurable to
user's needs. Included the event and context variables within the
handler call.